+++ remote.go
@@ -1,7 +1,6 @@
 package wormhole
 
 import (
-	"crypto/tls"
 	"net/url"
 	"os"
 	"os/signal"
@@ -13,6 +12,8 @@ import (
 	"github.com/superfly/wormhole/config"
 	wnet "github.com/superfly/wormhole/net"
 	handler "github.com/superfly/wormhole/remote"
+	"github.com/superfly/wormhole/session"
+	tlsc "github.com/superfly/wormhole/tls"
 )
 
 var (
@@ -25,28 +26,30 @@ func StartRemote(cfg *config.ServerConfig) {
 	log = cfg.Logger.WithFields(logrus.Fields{"prefix": "wormhole"})
 	ensureRemoteEnvironment(cfg)
 
+	registry := session.NewRegistry(cfg.Logger)
+
 	var h handler.Handler
 	var err error
 	server := &handler.Server{Logger: cfg.Logger}
 
-	listenerFactory, err := listenerFactoryFromConfig(cfg)
+	listenerFactory, err := listenerFactoryFromConfig(registry, cfg)
 	if err != nil {
 		log.Fatalf("Could not create listener factory: %+v", err)
 	}
 
 	switch cfg.Protocol {
 	case config.SSH:
-		h, err = handler.NewSSHHandler(cfg, redisPool, listenerFactory)
+		h, err = handler.NewSSHHandler(cfg, registry, redisPool, listenerFactory)
 		if err != nil {
 			log.Fatal(err)
 		}
 	case config.TCP:
-		h, err = handler.NewTCPHandler(cfg, redisPool, listenerFactory)
+		h, err = handler.NewTCPHandler(cfg, registry, redisPool, listenerFactory)
 		if err != nil {
 			log.Fatal(err)
 		}
 	case config.HTTP2:
-		h, err = handler.NewHTTP2Handler(cfg, redisPool, listenerFactory)
+		h, err = handler.NewHTTP2Handler(cfg, registry, redisPool, listenerFactory)
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -54,11 +57,11 @@ func StartRemote(cfg *config.ServerConfig) {
 		log.Fatal("Unknown wormhole transport layer protocol selected.")
 	}
 
-	go handleDeath(h)
+	go handleDeath(h, registry)
 	server.ListenAndServe(":"+cfg.Port, h)
 }
 
-func listenerFactoryFromConfig(cfg *config.ServerConfig) (wnet.ListenerFactory, error) {
+func listenerFactoryFromConfig(registry *session.Registry, cfg *config.ServerConfig) (wnet.ListenerFactory, error) {
 	multiPortFactoryArgs := &wnet.MultiPortTCPListenerFactoryArgs{
 		BindAddr: "0.0.0.0",
 		Logger:   cfg.Logger,
@@ -70,19 +73,15 @@ func listenerFactoryFromConfig(cfg *config.ServerConfig) (wnet.ListenerFactory, 
 	}
 
 	if cfg.UseSharedPortForwarding {
-		keyPair, err := tls.X509KeyPair(cfg.SharedPortTLSCert, cfg.SharedPortTLSPrivateKey)
+		tlsconf, err := tlsc.NewConfig(cfg.SharedPortTLSCert, cfg.SharedPortTLSPrivateKey, registry)
 		if err != nil {
 			return nil, err
-		}
-
-		tlsConf := &tls.Config{
-			Certificates: []tls.Certificate{keyPair},
 		}
 
 		sharedArgs := &wnet.SharedPortTLSListenerFactoryArgs{
 			Address:   ":" + cfg.SharedTLSForwardingPort,
 			Logger:    cfg.Logger,
-			TLSConfig: tlsConf,
+			TLSConfig: tlsconf.GetDefaultConfig(),
 		}
 		sharedL, err := wnet.NewSharedPortTLSListenerFactory(sharedArgs)
 		if err != nil {
@@ -160,13 +159,14 @@ func newRedisPool(redisURL string) *redis.Pool {
 }
 
 // IT CAN BE HANDLED!
-func handleDeath(h handler.Handler) {
+func handleDeath(h handler.Handler, r *session.Registry) {
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
 	go func(c <-chan os.Signal) {
 		for range c {
 			log.Print("Cleaning up before exit...")
 			h.Close()
+			r.Close()
 			log.Print("Cleaned up connections.")
 			os.Exit(1)
 		}